(bug 113)
restructuring and renaming for clarity, granularity, and reduction of code duplication

Found and commented on more errors in this code. In the "Close immediately" cases, it doesn't do a real close and that can break subsequent opening of other or the same problematic projects. 

Changed several cases where GetDirManager()->ProjectFSCK returned FSCKstatus_CHANGED. These caused a PushState, but Undo could not do anything. Changed corresponding wordings for "Replace missing audio with silence" to say "permanent immediately" instead of "permanent on save", which was incorrect.





git-svn-id: http://audacity.googlecode.com/svn/audacity-src/trunk@10605 47890b92-0858-11df-a26f-8b716316a5bc