Sync-Lock:
Remove tiles from rulers. They were interfering with visibility of ruler marks.

Restore the icon to TrackInfo, but not in the title bar. Instead, when sync-lock is on and the track IsSyncLockSelected(), decrease the width of the minimize button and show the icon to the right of it. This way, it shows even when the track is minimized. 

git-svn-id: http://audacity.googlecode.com/svn/audacity-src/trunk@10616 47890b92-0858-11df-a26f-8b716316a5bc